Create frontend structure
- Install vue-router
- Install pinia
- Create Project structure
