+++ src/components/Footer.vue
@@ -1,0 +1,15 @@
+<template>
+  <footer>
+    <p>© 2025 Sortifile</p>
+  </footer>
+</template>
+
+<script>
+export default {
+  name: "Footer",
+};
+</script>
+
+<style scoped>
+/* 頁尾樣式 */
+</style>
