+++ src/components/Navbar.vue
@@ -1,0 +1,18 @@
+<template>
+  <nav>
+    <ul>
+      <li><router-link to="/">Home</router-link></li>
+      <li><router-link to="/about">About</router-link></li>
+    </ul>
+  </nav>
+</template>
+
+<script>
+export default {
+  name: "Navbar",
+};
+</script>
+
+<style scoped>
+/* 導覽列樣式 */
+</style>
