+++ src/views/About.vue
@@ -1,0 +1,14 @@
+<!-- About.vue -->
+<!-- This is the temporary template file for adding a new page. -->
+
+<template>
+  <div>
+    <h1>About Us</h1>
+  </div>
+</template>
+
+<script setup></script>
+
+<style scoped>
+/* 當前頁面的樣式 */
+</style>
